Moved setPrice from scripts/dev to the scripts
